Renamed depends endpoint to relations and implemented both ancestors and descendants
